feat: Frontend cleanup
