+++ frontend/messages/de.json
@@ -22,6 +22,7 @@
 	"updating": "Aktualisieren",
 	"updatingDots": "Aktualisieren...",
 	"refresh": "Aktualisieren",
+	"details": "Details",
 	"sync": "Synchronisieren",
 	"syncing": "Synchronisiere...",
 	"rollback": "Rollback",
@@ -59,12 +60,17 @@
 	"noApplicationsFoundDescription": "Passe deine Suche an, um passende Anwendungen zu sehen.",
 	"noAgentsFound": "Keine Agenten gefunden",
 	"noAgentsFoundDescription": "Passe deine Suche an, um passende Agenten zu sehen.",
+	"noRepositoriesFound": "Keine Repositories gefunden",
+	"noRepositoriesFoundDescription": "Passe deine Suche an, um passende Repositories zu sehen.",
 	"totalAgentsCount": "Es gibt insgesamt {count} Agenten",
+	"totalRepositoriesCount": "Es gibt insgesamt {count} Repositories",
 	"appsCount": "Anzahl Apps",
 	"notAvailableShort": "k. A.",
 	"cardActions": "Kartenaktionen",
 	"deleteAgentCardTitle": "Agentenkarte löschen?",
 	"deleteAgentCardDescription": "Dadurch wird \"{name}\" dauerhaft gelöscht. Diese Aktion kann nicht rückgängig gemacht werden.",
+	"deleteRepositoryTitle": "Repository löschen?",
+	"deleteRepositoryDescription": "Dadurch wird \"{name}\" dauerhaft gelöscht. Diese Aktion kann nicht rückgängig gemacht werden.",
 	"statusOnline": "Online",
 	"statusOffline": "Offline",
 	"statusError": "Fehler",
