+++ frontend/messages/en.json
@@ -22,6 +22,7 @@
 	"updating": "Updating",
 	"updatingDots": "Updating...",
 	"refresh": "Refresh",
+	"details": "Details",
 	"sync": "Sync",
 	"syncing": "Syncing...",
 	"rollback": "Rollback",
@@ -59,12 +60,17 @@
 	"noApplicationsFoundDescription": "Adjust your search to see matching applications.",
 	"noAgentsFound": "No agents found",
 	"noAgentsFoundDescription": "Adjust your search to see matching agents.",
+	"noRepositoriesFound": "No repositories found",
+	"noRepositoriesFoundDescription": "Adjust your search to see matching repositories.",
 	"totalAgentsCount": "There are {count} agents in total",
+	"totalRepositoriesCount": "There are {count} repositories in total",
 	"appsCount": "Apps count",
 	"notAvailableShort": "n/a",
 	"cardActions": "Card actions",
 	"deleteAgentCardTitle": "Delete agent card?",
 	"deleteAgentCardDescription": "This will permanently delete \"{name}\". This action cannot be undone.",
+	"deleteRepositoryTitle": "Delete repository?",
+	"deleteRepositoryDescription": "This will permanently delete \"{name}\". This action cannot be undone.",
 	"statusOnline": "Online",
 	"statusOffline": "Offline",
 	"statusError": "Error",
